+++ instruction/instruction.go
@@ -30,7 +30,7 @@ type Instruction interface {
 	Name() string
 	Function() func(readers []io.Reader, writers []io.Writer, stats *Stats)
 	SerializeToCommand() *msg.Instruction
-	GetMemoryCostInMB(partitionSize int) int
+	GetMemoryCostInMB(partitionSize int64) int64
 }
 
 type instructionRunner struct {
