+++ instruction/partition_scatter.go
@@ -47,7 +47,7 @@ func (b *ScatterPartitions) SerializeToCommand() *msg.Instruction {
 	}
 }
 
-func (b *ScatterPartitions) GetMemoryCostInMB(partitionSize int) int {
+func (b *ScatterPartitions) GetMemoryCostInMB(partitionSize int64) int64 {
 	return 5
 }
